feat(examples)+docs: B2 剖面裁切带收编——E813 活体门+clippy 三修+mirror abi 钉 6（S4 收官）
- E813_section_clip.c 全路绿（ctest #17 Passed 0.10s；角域三面形拆前墙遮蔽，
  二面形 g2=0 探针先行 [PIT-8]，三面实测 1764 入区间 [800,4000]）
- 收尾三修：engine pick DVec3 直用（useless_conversion）/ffi_spec nonminimal_bool/
  contract_spec d=2.0e6 由 unused 升断言（n·(o+t) 恒等式再加一锁）
- web 面：WEB-BUILD ✓ · MIRROR 3/3（abi 钉 0x…0006 升账）· web_capi_pair_mirror ✓
- PIT-26 入档（from_raw_parts NULL+0 即 UB，debug 前判实锤，m>0 守卫律）
- 计数终账（同终端实测粘贴）：**191 passed · 130↔130 · GATE-ABI 26/26 ·
  GATE-DOCS ✓ E 23 件三方/README=130 · ctest 24 条 · CMAKE-SMOKE ✓ · ci 十段 0 红**
- status.md B2 基线段+非目标挂账；README/AGENTS 130/24/26 三口径同步

+++ bindings/c/visiaengine-capi/src/engine.rs
@@ -558,7 +558,7 @@ impl Engine {
             let hit = pick_meshes(ray, &cands)?;
             if let Some(c) = &self.clip {
                 let p = hit.point;
-                if !c.keeps([f64::from(p.x), f64::from(p.y), f64::from(p.z)]) {
+                if !c.keeps([p.x, p.y, p.z]) {
                     skip.push(hit.entity);
                     continue;
                 }
